Day 6 - Part 2 Upgraded return value of Exercise from int to long

+++ AdventOfCode2021/Day1.cs
@@ -4,14 +4,14 @@ namespace AdventOfCode2021
 {
     public class Day1 : Exercise
     {
-        public int ExecutePart1(string[] lines)
+        public long ExecutePart1(string[] lines)
         {
             var depths = lines.Select(int.Parse).ToList();
 
             return Enumerable.Range(1, lines.Length - 1).Count(i => depths[i] > depths[i - 1]);
         }
 
-        public int ExecutePart2(string[] lines)
+        public long ExecutePart2(string[] lines)
         {
             var depths = lines.Select(int.Parse).ToList();
