+++ AdventOfCode2021/Exercise.cs
@@ -2,7 +2,7 @@ namespace AdventOfCode2021
 {
     public interface Exercise
     {
-        int ExecutePart1(string[] lines);
-        int ExecutePart2(string[] lines);
+        long ExecutePart1(string[] lines);
+        long ExecutePart2(string[] lines);
     }
 }
